29/12/2016
I added the improved version of the control strategy to control one converter connected to grid. The idea is to allow the reference voltage at the output of the converter to change only depending on the voltage level of the grid. This gave the grid autonomous powers to choose when to receive, supply or shut down the converter when current/voltage values exceeded permissible levels